backend/requirements.txt
Flask==2.2.5
flask-cors==3.0.10
python-dotenv==1.0.0
mysql-connector-python==8.1.0
requests==2.31.0
